Enable config setting via environment variables
It is useful to be able to set the configuration properties via
environment variables, as well as / instead of via properties files.

Most operating systems disallow period-separated key names for
environment variables, so this commit enables using the underscore
format for environment variables in addition to the existing
period-separated format (for example, TESTRAIL_URL as well as
testrail.url).